+++ src/x_fileio.c
@@ -1,5 +1,4 @@
 #include "x_dg.h"
-#include <stdint.h>
 
 #define DLG_APPUNSAVED "dlgAppUnsaved"
 #define DLG_APP_UNSAVED "dlgAppUnsaved"
@@ -201,7 +200,7 @@ Widget OpenExportMeshDlg(View w,int bAtExit) {
     XmStringFree(xms);
   }
 
-  SetValues(wg,XmNuserData,(XtPointer)(uintptr_t)bAtExit,NULL);
+  SetValues(wg,XmNuserData,(XtPointer)bAtExit,NULL);
 
   return wg;
 }
@@ -278,7 +277,7 @@ Widget OpenFileSaveDlg(View w,int bAtExit) {
     XtPopup(XtParent(wg),XtGrabNone);
   }
 
-  SetValues(wg,XmNuserData,(XtPointer)(uintptr_t)bAtExit,NULL);
+  SetValues(wg,XmNuserData,(XtPointer)bAtExit,NULL);
 
   XmListDeselectAllItems(XmFileSelectionBoxGetChild(wg,XmDIALOG_LIST));
   if (w->app->fName!=NULL) {
@@ -370,7 +369,7 @@ static void CbExportMesh(Widget wg,XtPointer xtpV,XtPointer pcbs) {
     XtPopdown(XtParent(wg));
     SetViewMsg(w,GetResourceString(wg,"msgMeshSaved",NULL,NULL));
     UndoMark(w->app);
-    if ((uintptr_t)GetUserData(wg)) CloseXmView(w,-1);
+    if ((int)GetUserData(wg)) CloseXmView(w,-1);
   }
   XtFree(s);
 
@@ -458,7 +457,7 @@ static void CbFile_Save(Widget wg,View w,void* xtp) {
     UndoMark(w->app);
     AddXAppRecentFile(w->xapp,w->app->fName);
 
-    if ((uintptr_t)userData) CloseXmView(w,-1);
+    if ((int)userData) CloseXmView(w,-1);
   }
   XtFree(s);
 }
@@ -696,7 +695,7 @@ static void CbFilePrintOk(Widget wg,XtPointer xtpDlg,XtPointer pcbs) {
 
   s=XmTextGetString(dlg->wCommand);
 
-  switch((uintptr_t)GetOptionMenuValue(dlg->wCmdType)) {
+  switch((int)GetOptionMenuValue(dlg->wCmdType)) {
     case CMT_FILE:
       r=PsViewOutput(psW,s,-1,title);
       break;
@@ -902,7 +901,7 @@ static int EEAT_OkCheck(EEAT_Dlg dlg) {
 
 
 static Group EEAT_GetGroup(EEAT_Dlg dlg) {
-  switch((uintptr_t)GetOptionMenuValue(dlg->wOptMenu)) {
+  switch((int)GetOptionMenuValue(dlg->wOptMenu)) {
     case EEAT_ALL:
       return dlg->w->app->elems;
     case EEAT_MARKED:
