+++ src/x_menu.c
@@ -1,7 +1,6 @@
 /* Menu commands */
 
 #include "x_dg.h"
-#include <stdint.h>
 
 /* Menu callbacks
 */
@@ -242,7 +241,7 @@ void CbCmDelObjects(Widget wg,View w,void* pcbs) {
   SetActiveView(w);
 
   GetValues(wg,XmNuserData,&objType,NULL);
-  switch((uintptr_t)objType) {
+  switch((int)objType) {
     case T_MARKING:
       for (p=AppMark1st(w->app,&ix);p!=NULL;p=Next(&ix)) {
         switch (GetObjType(p)) {
